+++ src/app.js
@@ -1,5 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
+import AddOption from "./components/AddOption";
+import Option from "./components/Option"
 
 class IndecisionApp extends React.Component {
 
@@ -127,53 +129,6 @@ const Options = (props) => {
     );
 }
 
-const Option = (props) => {
-    return (
-        <div>
-            {props.optionText}
-            <button
-                onClick={(e) => {
-                    props.handleDeleteOption(props.optionText);
-                }}
-            >
-                remove
-            </button>
-        </div>
-    );
-}
 
-class AddOption extends React.Component {
-
-    constructor(props) {
-        super(props);
-        this.handleAddOption = this.handleAddOption.bind(this);
-        this.state = {
-            error: undefined
-        }
-    }
-
-    handleAddOption(e) {
-        e.preventDefault();
-        const option = e.target.elements.option.value.trim();
-        const error = this.props.handleAddOption(option);
-        this.setState({error});
-        if (!error) {
-            e.target.elements.option.value = '';
-        }
-    }
-
-    render() {
-        return (
-            <div id='add-option'>
-                {this.state.error && <p>{this.state.error}</p>}
-                <form onSubmit={this.handleAddOption}>
-                    <input type="text" name="option"></input>
-                    <button>Add Option</button>
-                </form>
-            </div>
-        );
-    }
-
-}
 
 ReactDOM.render(<IndecisionApp />, document.getElementById('app'))
