+++ src/components/AddOption.js
@@ -1,0 +1,35 @@
+import React from 'react';
+
+export default class AddOption extends React.Component {
+
+    constructor(props) {
+        super(props);
+        this.handleAddOption = this.handleAddOption.bind(this);
+        this.state = {
+            error: undefined
+        }
+    }
+
+    handleAddOption(e) {
+        e.preventDefault();
+        const option = e.target.elements.option.value.trim();
+        const error = this.props.handleAddOption(option);
+        this.setState({error});
+        if (!error) {
+            e.target.elements.option.value = '';
+        }
+    }
+
+    render() {
+        return (
+            <div id='add-option'>
+                {this.state.error && <p>{this.state.error}</p>}
+                <form onSubmit={this.handleAddOption}>
+                    <input type="text" name="option"></input>
+                    <button>Add Option</button>
+                </form>
+            </div>
+        );
+    }
+
+}
